move constants folders
