+++ src/Components/Home-Page/Layout/Community/index.css
@@ -101,4 +101,10 @@
     padding: 3rem;
     gap: 1rem;
   }
+
+  .community-team-btn {
+    font-size: 1.5rem;
+    font-weight: 400;
+    line-height: 100%;
+  }
 }
