+++ src/Constants/HomePage/Partners.js
