+++ src/Constants/HomePage/SocialMedia.js
@@ -1,0 +1,15 @@
+import FacebookIcon from "src/Assets/Icons/social-facebook.png";
+import LinkedinIcon from "src/Assets/Icons/social-linkedin.png";
+import TwitterIcon from "src/Assets/Icons/social-twitter.png";
+import InstagramIcon from "src/Assets/Icons/social-instagram.png";
+import YoutubeIcon from "src/Assets/Icons/social-youtube.png";
+
+const SocialMedia = [
+  { icon: FacebookIcon, url: "#" },
+  { icon: LinkedinIcon, url: "#" },
+  { icon: TwitterIcon, url: "#" },
+  { icon: InstagramIcon, url: "#" },
+  { icon: YoutubeIcon, url: "#" },
+];
+
+export default SocialMedia;
