--- src/Constants/SocialMedia.js
@@ -1,15 +1,0 @@
-import FacebookIcon from "../Assets/Icons/social-facebook.png";
-import LinkedinIcon from "../Assets/Icons/social-linkedin.png";
-import TwitterIcon from "../Assets/Icons/social-twitter.png";
-import InstagramIcon from "../Assets/Icons/social-instagram.png";
-import YoutubeIcon from "../Assets/Icons/social-youtube.png";
-
-const SocialMedia = [
-  { icon: FacebookIcon, url: "#" },
-  { icon: LinkedinIcon, url: "#" },
-  { icon: TwitterIcon, url: "#" },
-  { icon: InstagramIcon, url: "#" },
-  { icon: YoutubeIcon, url: "#" },
-];
-
-export default SocialMedia;
